Added some logic for roles, like dynamic role hierarchy. Finished with roles. Added several exception handler methods in GlobalExceptionHandler. Also some minor fixes. Now it works, but still need some upgrades.

+++ src/main/java/com/example/blogging/service/impl/RoleServiceImpl.java
@@ -4,14 +4,22 @@ import com.example.blogging.domain.Role;
 import com.example.blogging.dto.role.RoleRequestDto;
 import com.example.blogging.repository.RoleRepository;
 import com.example.blogging.repository.entity.RoleEntity;
+import com.example.blogging.security.DynamicRoleHierarchy;
+import com.example.blogging.security.service.RoleHierarchyService;
 import com.example.blogging.service.RoleService;
 import com.example.blogging.service.exception.conflict.impl.RoleAlreadyExistsException;
 import com.example.blogging.service.exception.notFound.impl.RoleNotFoundException;
 import com.example.blogging.service.mapper.RoleEntityMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.web.access.expression.DefaultWebSecurityExpressionHandler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionSynchronization;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
 
+import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -20,9 +28,12 @@ import java.util.stream.Collectors;
 public class RoleServiceImpl implements RoleService {
     private final RoleRepository roleRepository;
     private final RoleEntityMapper roleEntityMapper;
+    private final DynamicRoleHierarchy dynamicRoleHierarchy;
+    private final RoleHierarchyService roleHierarchyService;
 
     @Override
     @Transactional(readOnly = true)
+    @PreAuthorize("hasAnyRole('ADMIN')")
     public Set<Role> getAllRoles() {
         return roleRepository.findAll().stream()
                 .map(roleEntityMapper::toRole)
@@ -31,6 +42,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional(readOnly = true)
+    @PreAuthorize("hasAnyRole('ADMIN')")
     public Role getRoleById(Long roleId) {
         RoleEntity roleEntity = roleRepository.findById(roleId)
                 .orElseThrow(() -> new RoleNotFoundException(roleId));
@@ -39,6 +51,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional(readOnly = true)
+    @PreAuthorize("hasAnyRole('ADMIN')")
     public Role getRoleByName(String roleName) {
         RoleEntity roleEntity = roleRepository.findByName(roleName)
                 .orElseThrow(() -> new RoleNotFoundException(roleName));
@@ -47,9 +60,17 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional
+    @PreAuthorize("hasAnyRole('ADMIN')")
     public Role createRole(RoleRequestDto roleRequestDto) {
+        if (!roleRepository.existsById(roleRequestDto.getParentId())) {
+            throw new RoleNotFoundException(roleRequestDto.getParentId());
+        }
+
+        RoleEntity parentRoleEntity = roleRepository.getReferenceById(roleRequestDto.getParentId());
+
         Role role = Role.builder()
-                .name(Role.ROLE_PREFIX + roleRequestDto.getName())
+                .name(Role.ROLE_PREFIX + roleRequestDto.getName().toUpperCase())
+                .parent(roleEntityMapper.toRole(parentRoleEntity))
                 .build();
 
         if (roleRepository.existsByName(role.getName())) {
@@ -59,12 +80,41 @@ public class RoleServiceImpl implements RoleService {
         RoleEntity roleEntity = roleEntityMapper.toRoleEntity(role);
         roleEntity = roleRepository.saveAndFlush(roleEntity);
 
+        registerHierarchyRefreshSynchronization();
+
         return roleEntityMapper.toRole(roleEntity);
     }
 
     @Override
     @Transactional
+    @PreAuthorize("hasAnyRole('ADMIN')")
     public void deleteRoleById(Long roleId) {
+        RoleEntity roleToDelete = roleRepository.findById(roleId)
+                .orElse(null);
+
+        if (roleToDelete == null) {
+            return;
+        }
+
+        List<RoleEntity> childrenRoles = roleRepository.findAllByParent(roleToDelete);
+
+        if (!childrenRoles.isEmpty()) {
+            RoleEntity parentRole = roleToDelete.getParent();
+            for (RoleEntity childRoleEntity : childrenRoles) {
+                childRoleEntity.setParent(parentRole);
+            }
+        }
+
         roleRepository.deleteById(roleId);
+        registerHierarchyRefreshSynchronization();
+    }
+
+    private void registerHierarchyRefreshSynchronization() {
+        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
+            @Override
+            public void afterCommit() {
+                dynamicRoleHierarchy.updateHierarchy(roleHierarchyService.getHierarchy());
+            }
+        });
     }
 }
